Creating an Editor tool that Selects Object By Name
This is an Editor tool that allows you to quickly search and select scripts and objects by name. The benefit of this tool is that when you have searched for your objects and they are selected you can quickly and efficiently add new game objects to all the selected objects saving time.

So for example you have created a new scene and have many game objects planes for the ground and they are all located in different areas in the hierarchy. Your task is to add box colliders to all of them instead of manually going into all of them and adding them this tool allows you to search for the object by name which then makes the adding of a component an easy task. This tool is not limited to only adding box colliders to game objects.

In this commit, I have created three versions of the unity editor tool they all work in similar ways however they do vary in some ways.

In the next commit, I am going to try to make the search function, more powerful by adding conditions, that it would ignore numbers inside curly brackets. The idea is that it would only ignore things like numbers inside curly brackets when the checkbox for this function would be ticked in the editor.